feat(ModelList): move refresh button to page header
- Move refresh action from ControlPanel to page header
- Remove redundant refresh props from ControlPanel
- Update related tests to reflect changes
