+++ src/features/ModelList/sortModes.ts
@@ -7,10 +7,3 @@ export const MODEL_LIST_SORT_MODES = {
 
 export type ModelListSortMode =
   (typeof MODEL_LIST_SORT_MODES)[keyof typeof MODEL_LIST_SORT_MODES]
-
-/**
- *
- */
-export function isModelCheapestFirstSortMode(mode: ModelListSortMode) {
-  return mode === MODEL_LIST_SORT_MODES.MODEL_CHEAPEST_FIRST
-}
